Move more annotation-related functions into the annotation directory
