+++ metaspace/engine/scripts/update_es_index.py
@@ -5,7 +5,7 @@ from functools import partial
 
 from sm.engine.db import DB
 from sm.engine.es_export import ESExporter, ESIndexManager
-from sm.engine.isocalc_wrapper import IsocalcWrapper
+from sm.engine.annotation.isocalc_wrapper import IsocalcWrapper
 from sm.engine.util import bootstrap_and_run
 
 logger = logging.getLogger('engine')
