+++ metaspace/engine/sm/engine/daemons/dataset_manager.py
@@ -20,7 +20,7 @@ from sm.engine.postprocessing.ion_thumbnail import (
     generate_ion_thumbnail,
     generate_ion_thumbnail_lithops,
 )
-from sm.engine.isocalc_wrapper import IsocalcWrapper
+from sm.engine.annotation.isocalc_wrapper import IsocalcWrapper
 from sm.engine.postprocessing.off_sample_wrapper import classify_dataset_ion_images
 from sm.engine.optical_image import del_optical_image
 from sm.engine.config import SMConfig
