+++ metaspace/engine/sm/engine/es_export.py
@@ -16,9 +16,9 @@ from elasticsearch.helpers import parallel_bulk
 
 from sm.engine.utils.db_mutex import DBMutex
 from sm.engine.db import DB
-from sm.engine.fdr import FDR
+from sm.engine.annotation.fdr import FDR
 from sm.engine.formula_parser import format_ion_formula
-from sm.engine.isocalc_wrapper import IsocalcWrapper
+from sm.engine.annotation.isocalc_wrapper import IsocalcWrapper
 from sm.engine import molecular_db
 from sm.engine.molecular_db import MolecularDB
 from sm.engine.utils.retry_on_exception import retry_on_exception
